test(architecture): move pricing http api tests
